Needed to merge

+++ queue.c
@@ -2,7 +2,7 @@
 
 //Dr. Sean Carroll's Queue Source
 
-void init_queue(queue_t * ptr_to_existing_q, int max_items)
+void init_queue(queue_t * ptr_to_existing_q, uint16_t max_items)
 {
     int16_t * new_buff = (int16_t *const) malloc((max_items+1) * sizeof(int16_t));
     ptr_to_existing_q ->buffer = new_buff;
